requirements.txt
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
python-dotenv==1.0.1
rank_bm25==0.2.2
sentence-transformers==3.0.1